app/requirements.txt
flask
requests
googletrans==4.0.0-rc1
python-dotenv
boto3

